boolean fields

+++ src/Ems/Database/migrations/2024_11_07_144932_create_email_statistics_table.php
@@ -12,10 +12,10 @@ return new class extends Migration
             $table->id();
             $table->foreignId('email_id')->constrained()->cascadeOnDelete();
             $table->foreignId('crm_card_id')->constrained()->cascadeOnDelete();
-            $table->integer('send')->default(0);
-            $table->integer('bounced')->default(0);
-            $table->integer('opened')->default(0);
-            $table->integer('clicked')->default(0);
+            $table->smallInteger('send')->default(0);
+            $table->smallInteger('bounced')->default(0);
+            $table->smallInteger('opened')->default(0);
+            $table->smallInteger('clicked')->default(0);
             $table->timestamps();
 
             $table->unique(['email_id', 'crm_card_id'], 'email_statistics_crm_card_id');
